# SkilloraMarketplace/tut5/openapi.yaml
openapi: 3.0.3
info:
  title: Skillora Orders Service
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.skillora.com
    description: Production

security:
  - bearerAuth: []

paths:
  /orders:
    post:
      summary: Create an order
      operationId: createOrder
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Authorization'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}

    get:
      summary: List orders
      operationId: listOrders
      parameters:
        - in: query
          name: customer
          schema: {type: string}
        - in: query
          name: status
          schema: {type: string}
        - $ref: '#/components/parameters/Authorization'
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
        '401': {$ref: '#/components/responses/Unauthorized'}

  /orders/{orderId}:
    get:
      summary: Get one order
      operationId: getOrder
      parameters:
        - in: path
          name: orderId
          required: true
          schema: {type: string}
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Order
          headers:
            ETag: {schema: {type: string}}
            Cache-Control: {schema: {type: string}}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '304': {description: Not Modified}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}

    patch:
      summary: Update order
      operationId: updateOrder
      parameters:
        - in: path
          name: orderId
          required: true
          schema: {type: string}
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {description: Precondition Failed}
        '422': {$ref: '#/components/responses/ValidationError'}

  /orders/{orderId}/cancellation:
    post:
      summary: Cancel an order
      operationId: cancelOrder
      parameters:
        - in: path
          name: orderId
          required: true
          schema: {type: string}
        - $ref: '#/components/parameters/Authorization'
      responses:
        '200': {description: Cancelled}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema: {type: string}
    Authorization:
      in: header
      name: Authorization
      required: true
      schema: {type: string}
    IfNoneMatch:
      in: header
      name: If-None-Match
      required: false
      schema: {type: string}
    IfMatch:
      in: header
      name: If-Match
      required: false
      schema: {type: string}

  schemas:
    CreateOrderRequest:
      type: object
      required: [customerId, address, items]
      properties:
        customerId: {type: string}
        address: {type: string}
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [serviceId, qty]
            properties:
              serviceId: {type: string}
              qty: {type: integer, minimum: 1}

    UpdateOrderRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]

    OrderResponse:
      type: object
      properties:
        id: {type: string}
        customerId: {type: string}
        address: {type: string}
        items:
          type: array
          items:
            type: object
            properties:
              serviceId: {type: string}
              qty: {type: integer}
              price: {type: number}
        status: {type: string}
        total: {type: number}
        version: {type: integer}

    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        errors:
          type: array
          items: {type: string}

  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unauthorized:
      description: Missing or empty authorization
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Conflict:
      description: State conflict
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    ValidationError:
      description: Domain validation failure
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    RateLimited:
      description: Too many requests
      headers:
        Retry-After: {schema: {type: integer}}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unavailable:
      description: Dependency unavailable
      headers:
        Retry-After: {schema: {type: integer}}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}